App layout added, enhancements
