+++ src/Main.tsx
@@ -17,37 +17,23 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-import { Box } from '@mui/material';
 import { ErrorPage } from 'cx-portal-shared-components';
-import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Outlet } from 'react-router-dom';
 
-import Nav from './components/Nav';
-import Sidebar from './components/Sidebar';
+import AppLayout from './components/layouts/AppLayout';
+import { IUser } from './models/User';
 import UserService from './services/UserService';
 
-export default function Main() {
+export default function Main({ loggedUser }: { loggedUser: IUser }) {
   document.title = useTranslation().t('title');
-  const [isExpanded, setIsExpanded] = useState(true);
-  const handleExpanded = (expanded: boolean) => {
-    setIsExpanded(expanded);
-  };
 
   return (
     <>
       {UserService.hasValidResource() ? (
-        <Box sx={{ my: 0, mx: 'auto', overflowY: 'auto', overflowX: 'hidden', height: '100vh' }}>
-          <Box>
-            <Nav getIsExpanded={(expanded: boolean) => handleExpanded(expanded)} />
-            <Box sx={{ display: 'flex', mt: 8, height: 'calc(100vh - 64px)', overflow: 'hidden' }}>
-              <Sidebar isExpanded={isExpanded} />
-              <Box sx={{ width: '100%', height: '100%', overflowY: 'scroll' }}>
-                <Outlet />
-              </Box>
-            </Box>
-          </Box>
-        </Box>
+        <AppLayout loggedUser={loggedUser}>
+          <Outlet />
+        </AppLayout>
       ) : (
         <ErrorPage header="This webpage is not available." title="Sorry for this inconvenience." />
       )}
